+++ src/OpenTelemetry/Trace/Samplers/AlwaysOffSampler.cs
@@ -24,7 +24,7 @@ namespace OpenTelemetry.Trace.Samplers
         /// <inheritdoc />
         public override SamplingResult ShouldSample(in SamplingParameters samplingParameters)
         {
-            return new SamplingResult(Decision.NotRecord);
+            return new SamplingResult(SamplingDecision.NotRecord);
         }
     }
 }