+++ src/OpenTelemetry/Trace/Samplers/SamplingResult.cs
@@ -13,10 +13,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System.Collections.Generic;
 using System.Linq;
 
-namespace OpenTelemetry.Trace
+namespace OpenTelemetry.Trace.Samplers
 {
     /// <summary>
     /// Sampling decision.
@@ -27,9 +28,19 @@ namespace OpenTelemetry.Trace
         /// Initializes a new instance of the <see cref="SamplingResult"/> struct.
         /// </summary>
         /// <param name="decision"> indicates whether an activity object is recorded and sampled.</param>
-        public SamplingResult(Decision decision)
+        public SamplingResult(SamplingDecision decision)
         {
             this.Decision = decision;
+            this.Attributes = Enumerable.Empty<KeyValuePair<string, object>>();
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SamplingResult"/> struct.
+        /// </summary>
+        /// <param name="isSampled"> True if sampled, false otherwise.</param>
+        public SamplingResult(bool isSampled)
+        {
+            this.Decision = isSampled ? SamplingDecision.RecordAndSampled : SamplingDecision.NotRecord;
             this.Attributes = Enumerable.Empty<KeyValuePair<string, object>>();
         }
 
@@ -39,7 +50,7 @@ namespace OpenTelemetry.Trace
         /// <param name="decision">indicates whether an activity object is recorded and sampled.</param>
         /// <param name="attributes">Attributes associated with the sampling decision. Attributes list passed to
         /// this method must be immutable. Mutations of the collection and/or attribute values may lead to unexpected behavior.</param>
-        public SamplingResult(Decision decision, IEnumerable<KeyValuePair<string, object>> attributes)
+        public SamplingResult(SamplingDecision decision, IEnumerable<KeyValuePair<string, object>> attributes)
         {
             this.Decision = decision;
 
@@ -52,7 +63,7 @@ namespace OpenTelemetry.Trace
         /// <summary>
         /// Gets a value indicating indicates whether an activity object is recorded and sampled.
         /// </summary>
-        public Decision Decision { get; }
+        public SamplingDecision Decision { get; }
 
         /// <summary>
         /// Gets a map of attributes associated with the sampling decision.