+++ test/OpenTelemetry.Tests/Implementation/Trace/Samplers/ProbabilitySamplerTest.cs
@@ -65,7 +65,7 @@ namespace OpenTelemetry.Trace.Samplers.Test
                       0,
                     });
             Assert.Equal(
-                Decision.NotRecord,
+                SamplingDecision.NotRecord,
                 defaultProbability.ShouldSample(new SamplingParameters(default, notSampledtraceId, ActivityDisplayName, ActivityKindServer, null, null)).Decision);
 
             // This traceId will be sampled by the ProbabilitySampler because the first 8 bytes as long
@@ -92,7 +92,7 @@ namespace OpenTelemetry.Trace.Samplers.Test
                       0,
                     });
             Assert.Equal(
-                Decision.RecordAndSampled,
+                SamplingDecision.RecordAndSampled,
                 defaultProbability.ShouldSample(new SamplingParameters(default, sampledtraceId, ActivityDisplayName, ActivityKindServer, null, null)).Decision);
         }